[fix]スペシャルアタックUIの見た目を変更しました。
