+++ Assets/Scripts/UserInterface/Utility/TextureMask.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UIElements;
-using UnityEditor.UIElements;
 
 
 [UxmlElement]
@@ -68,25 +67,5 @@ public partial class TextureMask : VisualElement
 
         mesh.SetAllVertices(vertex);
         mesh.SetAllIndices(triangles.ToArray());
-    }
-
-    //TODO:HERE たぶんビルドできない
-    public class VertexPositionDatas : UxmlAttributeConverter<Vector2[]>
-    {
-        public override Vector2[] FromString(string value)
-        {
-            var sprit = value.Replace(" ", "").Replace("(", "").Replace(")", "").Split('|');
-            Vector2[] output = new Vector2[sprit.Length];
-            for (int i = 0; i < sprit.Length; i++)
-            {
-                var n = Array.ConvertAll(sprit[i].Trim(')').Split(','), float.Parse);
-                output[i] = new Vector2(n[0], n[1]);
-            }
-
-            return output;
-        }
-
-        public override string ToString(Vector2[] value) =>
-            System.FormattableString.Invariant($"{string.Join('|', value)}");
     }
 }
